renamed from SpTriplet to SparseRowMaj
